Add scrollSpy functionality
