+++ content/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
-import MainPage from './components/Main/MainPage.container';
+import MainPage from './components/MainPage/MainPage.container';
 
 function App() {
   return (
