--- content/src/components/ContentWrapper/queries.js
@@ -1,7 +1,0 @@
-import gql from "graphql-tag";
-
-export const CONTENT_QUERY = gql`
-  query Content($contentType: String!, $id: String!) {
-    content(contentType: $contentType, id: $id)
-  }
-`;
