+++ content/src/components/MainPage/styled.js
