+++ content/src/components/Navigation/LeftNavigation/styled.js
